Test for issue #11
This commit tests the impact of the changes suggested by issue #11. The
results indicate that using the google maps locations will incur a
higher number of traceroutes because Google Maps uses different city
names for suburbs of the same metropolitan area, therefore it increases
the number of candidate cities where we need probes.