Changes in probe listing for improved performance
Change the way active Atlas probe per location and per ASN are listed in
order to improve the performance when a large list of IPs is passed as
query input. Now all the active atlas probes are listed once at the
beginning of the script, instead of obtaining a fresh list for every IP
address. The downside is that some Atlas probes may change state during
the program execution but the number should be low, so the trade-off is
favourable.